requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
pandas==1.3.3
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==3.0.4
Django==3.2.5
django-filter==21.1
djangorestframework==3.12.4
greenlet==1.1.2
idna==2.8
importlib-metadata==4.8.1
lxml==4.6.4
mkl-fft
mkl-random
mkl-service
more-itertools==8.10.0
numexpr==2.7.3
packaging==21.0
pip==21.2.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==4.6.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.22.0
setuptools==58.1.0
six==1.16.0
sleeper-api-wrapper==1.0.7
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.25.11
wcwidth==0.2.5
wheel==0.37.0
zipp==3.6.0
uWSGI>=2.0.19.1,<2.1
